pyproject.toml
[tool.poetry]
name = "enginesdk"
version = "0.1.0"
description = "AI Engine SDK"
authors = ["yann <yann.leguilly@gmail.com>", "Trouni Tiet <trouni@gmail.com>"]
packages = [{ include = "enginesdk" }]

[tool.poetry.dependencies]
python = "3.7.9"
fastapi = "^0.63.0"
pydantic = "^1.7.3"
PyYAML = "^5.4.1"
google-cloud-core = "^1.5.0"
google-cloud-secret-manager = "1.0.0"
google-cloud-storage = "^1.35.0"
google-cloud-logging = "^2.2.0"
cloudinary = "^1.24.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.6.4"
pytest-cov = "^2.10.1"
flake8 = "^3.8.4"
pylint-fail-under = "^0.3.0"
coverage-badge = "^1.0.1"
pytest = "^6.2.1"

[build-system]
requires = ["poetry_core>=1.0"]
build-backend = "poetry.core.masonry.api"
